+++ test/plugin-multi-tenant/e2e.spec.ts
@@ -8,6 +8,9 @@ import { fileURLToPath } from 'url'
 import type { PayloadTestSDK } from '../__helpers/shared/sdk/index.js'
 import type { Config } from './payload-types.js'
 
+import { loginClientSide } from '../__helpers/e2e/auth/login.js'
+import { openRelationshipFieldDrawer } from '../__helpers/e2e/fields/relationship/openRelationshipFieldDrawer.js'
+import { goToListDoc } from '../__helpers/e2e/goToListDoc.js'
 import {
   changeLocale,
   ensureCompilationIsDone,
@@ -15,10 +18,6 @@ import {
   saveDocAndAssert,
   waitForFormReady,
 } from '../__helpers/e2e/helpers.js'
-import { AdminUrlUtil } from '../__helpers/shared/adminUrlUtil.js'
-import { loginClientSide } from '../__helpers/e2e/auth/login.js'
-import { openRelationshipFieldDrawer } from '../__helpers/e2e/fields/relationship/openRelationshipFieldDrawer.js'
-import { goToListDoc } from '../__helpers/e2e/goToListDoc.js'
 import {
   clearSelectInput,
   getSelectInputOptions,
@@ -26,8 +25,9 @@ import {
   selectInput,
 } from '../__helpers/e2e/selectInput.js'
 import { closeNav, openNav } from '../__helpers/e2e/toggleNav.js'
-import { initPayloadE2ENoConfig } from '../__helpers/shared/initPayloadE2ENoConfig.js'
+import { AdminUrlUtil } from '../__helpers/shared/adminUrlUtil.js'
 import { reInitializeDB } from '../__helpers/shared/clearAndSeed/reInitializeDB.js'
+import { initPayloadE2ENoConfig } from '../__helpers/shared/initPayloadE2ENoConfig.js'
 import { TEST_TIMEOUT_LONG } from '../playwright.config.js'
 import { credentials } from './credentials.js'
 import { autosaveGlobalSlug, menuItemsSlug, menuSlug, tenantsSlug, usersSlug } from './shared.js'
@@ -883,6 +883,31 @@ test.describe('Multi Tenant', () => {
           })
         })
         .toBeFalsy()
+    })
+
+    test('should populate tenant selector after standard form login with tree-restructuring provider', async () => {
+      // This test verifies the fix for a bug where TenantSelectionProviderClient
+      // loses state on remount. The ConditionalWrapperProvider (registered in the
+      // test config) changes its tree structure when the user authenticates
+      // (DummyContext.Provider → Fragment), causing React to remount the subtree.
+      // Without the fix, useState(initialTenantOptions) re-initializes with stale
+      // RSC props and useRef(userID) re-initializes to the current userID, so
+      // userChanged is false and syncTenants never fires.
+
+      // Use the "owner" user who has access to multiple tenants (Blue Dog + Anchor Bar).
+      // The tenant selector only renders when options.length > 1.
+      await loginClientSide({
+        data: credentials.owner,
+        page,
+        serverURL,
+      })
+
+      // After login, the tenant selector should be populated with the owner's tenants
+      await expect
+        .poll(async () => {
+          return (await getTenantOptions({ page })).sort()
+        })
+        .toEqual(['Anchor Bar', 'Blue Dog'].sort())
     })
   })
 })
